修改代码

+++ core/foundation/event.go
@@ -1,12 +1,12 @@
 package foundation
 
 type Hook interface {
-	GetHookID() uintptr
+	GetHookID() uint64
 	GetPriority() int
 }
 
 type Event interface {
 	AddHook(hook Hook)
-	RemoveHook(hookID uintptr)
+	RemoveHook(hookID uint64)
 	RangeHooks(fun func(hook Hook) bool)
 }
